requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
pydantic
requests
python-slugify

tavily-python
langchain-tavily

langchain>=0.1.0
langchain-community>=0.1.0
langchain-google-genai
google-generativeai

passlib[bcrypt]
python-multipart
pyjwt
sqlalchemy
bcrypt
pytest
httpx


langchain-groq
tiktoken
watchdog

gunicorn

langgraph
